Moved modally presentation to the core
